+++ snippets/functions.rs
@@ -1,9 +1,9 @@
 use compose_idents::compose_idents;
 
 compose_idents!(
-    my_const = [upper(foo), _, lower(BAR)],
+    my_const = concat(upper(foo), _, lower(BAR)),
     // Function calls can be arbitrarily nested and combined.
-    my_static = [upper(lower(BAZ))],
+    my_static = upper(lower(BAZ)),
     {
         const my_const: u8 = 1;
         static my_static: &str = "hello";