+++ snippets/normalize.rs
@@ -1,7 +1,7 @@
 use compose_idents::compose_idents;
 
 compose_idents!(
-    MY_NORMALIZED_ALIAS = [my, _, normalize(&'static str)],
+    MY_NORMALIZED_ALIAS = concat(my, _, normalize(&'static str)),
     {
         static MY_NORMALIZED_ALIAS: &str = "This alias is made from a normalized argument";
     }