+++ tests/compile/funcs/pascal_case.rs
@@ -1,7 +1,7 @@
 use compose_idents::compose_idents;
 
 fn main() {
-    compose_idents!(my_fn = [pascal_case(foo_bar)], {
+    compose_idents!(my_fn = pascal_case(foo_bar), {
         fn my_fn() -> u32 {
             42
         }